Fix decoder bitstream-desync and dequant bugs in AV1, H.264, and AAC
AV1 reconstruction (reconstruct.rs):
- Drop the fabricated tile_cdf_update_flag read for inter-frame tile groups
- Use correct cfl_allowed gating per block size instead of a fixed `true`
- Read filter_intra_mode_info on the inter intra-block path
- Remove incorrect 4x4 early-out and the over-subtract in read_tx_size
- Fix chroma plane copy bounds with div_ceil and add filter-intra prediction

H.264 transform (transform.rs): correct 8x8 dequant to FFmpeg's shift
formula (was dividing by an extra 2^6) and update affected tests.

AAC decoder (decoder.rs + modules): borrow-checker fixes, copy_from_slice
replacements, and dead_code annotations for unused scaffolding fields.

Also adds an AV1 smptebars debug test harness.

+++ tpt-kinetix-kg/src/extraction.rs
@@ -62,7 +62,7 @@ fn function_name<'a>(func_def: Node<'a>, source: &'a [u8]) -> Option<&'a str> {
 /// Return the text of the first named child of `node` that is an identifier.
 fn first_identifier_text<'a>(node: Node<'a>, source: &'a [u8]) -> Option<&'a str> {
     for i in 0..node.named_child_count() {
-        if let Some(child) = node.named_child(i) {
+        if let Some(child) = node.named_child(i as u32) {
             if child.kind() == "identifier" || child.kind() == "number_literal" {
                 return child.utf8_text(source).ok();
             }
